Clear the project ID from user information
Currently when a project is deleted, the project ID details
still exists in user information. After this fix, when a project
is deleted the default project ID in user
information will be cleared.

Closes-Bug: #1523369

+++ keystone/tests/unit/resource/test_backends.py
@@ -1505,8 +1505,8 @@ class ResourceTests(object):
         project_ref = self.resource_api.get_project(project['id'])
         self.assertDictEqual(updated_project_ref, project_ref)
 
-    @test_utils.wip('waiting for fix to bug #1523369')
     def test_delete_project_clears_default_project_id(self):
+        self.config_fixture.config(group='cache', enabled=False)
         project = unit.new_project_ref(
             domain_id=CONF.identity.default_domain_id)
         user = unit.new_user_ref(domain_id=CONF.identity.default_domain_id,
@@ -1514,13 +1514,14 @@ class ResourceTests(object):
         self.resource_api.create_project(project['id'], project)
         user = self.identity_api.create_user(user)
         user = self.identity_api.get_user(user['id'])
-        self.assertIsNotNone(user['default_project_id'])
 
-        self.resource_api.delete_project(project['id'])
-        user = self.identity_api.get_user(user['id'])
-        self.assertIsNone(user['default_project_id'])
+        # LDAP is read only default_project_id doesn't exist
+        if 'default_project_id' in user:
+            self.assertIsNotNone(user['default_project_id'])
+            self.resource_api.delete_project(project['id'])
+            user = self.identity_api.get_user(user['id'])
+            self.assertNotIn('default_project_id', user)
 
-    @test_utils.wip('waiting for fix to bug #1523369')
     def test_delete_project_with_roles_clears_default_project_id(self):
         project = unit.new_project_ref(
             domain_id=CONF.identity.default_domain_id)
@@ -1533,10 +1534,9 @@ class ResourceTests(object):
         self.assignment_api.create_grant(user_id=user['id'],
                                          project_id=project['id'],
                                          role_id=role['id'])
-
         self.resource_api.delete_project(project['id'])
         user = self.identity_api.get_user(user['id'])
-        self.assertIsNone(user['default_project_id'])
+        self.assertNotIn('default_project_id', user)
 
 
 class ResourceDriverTests(object):
